fix(billing): extract src-lib for cross-plugin consumption
