+++ libs/portal-plugin-billing/src-lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useSubscriptionStatus, getAuthHeaders, SUBSCRIPTION_QUERY_KEY } from "./useSubscriptionStatus";
+export type { UseSubscriptionStatusReturn } from "./useSubscriptionStatus";
