+++ libs/portal-plugin-billing/src-lib/types/index.ts
@@ -1,0 +1,16 @@
+export type { SubscriptionStatusResponse } from "@lumeweb/portal-sdk";
+
+// SSE Event Types — must match backend SSEEventType constants in portal-plugin-billing
+export {
+  BillingSSEEventType,
+  type BillingSSEEventDataMap,
+  type PaymentCompletedEventData,
+  type SubscriptionActiveEventData,
+  type SubscriptionCreatedEventData,
+  type SubscriptionUpdatedEventData,
+  type SubscriptionCancelledEventData,
+  type PlanChangedEventData,
+  type PlanChangeCreditOnlyEventData,
+  type PlanChangeZeroAmountEventData,
+  type BillingSSEEvent,
+} from "./subscription";
