+++ libs/portal-plugin-billing/tsconfig.json
@@ -2,7 +2,8 @@
   "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@lib/*": ["./src-lib/*"]
     }
   },
   "exclude": ["vitest.config.ts"]
